+++ admin/vymazanie/vymazanie_produktu.php
@@ -14,10 +14,11 @@ if (!isset($_GET['id']) || !is_numeric($_GET['id'])) {
     $produkt_cesta = $produkt->vypisJednehoZaznamu($_GET['id']);
     $absolutna_cesta = $_SERVER['DOCUMENT_ROOT']. '/FitStream/' . $produkt_cesta['img_hlavna']; 
 
-    if (!unlink($absolutna_cesta)) {
-        die("Súbor neexistuje");
+    if(file_exists($absolutna_cesta)){
+        if (!unlink($absolutna_cesta)) {
+            die("Súbor neexistuje");
+        }
     }
-  
 
    $produkt->vymazanieRiadku($_GET['id']);
 }
