+++ admin/vymazanie/vymazanie_slideshow.php
@@ -13,8 +13,10 @@ if (!isset($_GET['id']) || !is_numeric($_GET['id'])) {
     $slideshow_img_cesta = $slideshow->vypisjednehoZaznamu($_GET['id']);
     $cesta = $_SERVER['DOCUMENT_ROOT']. '/FitStream/' . $slideshow_img_cesta['img_url']; 
 
-    if (!unlink($cesta)) {
-        die("Súbor neexistuje");
+   if(file_exists($cesta)){
+        if (!unlink($cesta)) {
+            die("Súbor neexistuje");
+        }
     }
     
     $slideshow->vymazanieRiadku($_GET['id']);
